+++ src/main/java/com/example/dao/NumbersOperationsDAO.java
@@ -4,19 +4,19 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public class NumbersOperationsDAO {
-	public int getSum(int a, int b) {
+	public double getSum(double a, double b) {
 		return a+b;
 	}
 
-	public int getDifference(int a, int b) {
+	public double getDifference(double a, double b) {
 		return a-b;
 	}
 
-	public int getProduct(int a, int b) {
+	public double getProduct(double a, double b) {
 		return a*b;
 	}
 
-	public double getDivision(int a, int b) {
+	public double getDivision(double a, double b) {
 		if(b!=0)
 			return a/b;
 		return -99999;
